pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terminal-game-mud"
version = "0.1.0"
description = "Um jogo multiplayer tipo MUD (Multi-User Dungeon) para terminal Linux"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "OpenMud Contributors"}
]
keywords = ["mud", "game", "terminal", "multiplayer", "rpg", "text-based"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment",
    "Topic :: Games/Entertainment :: Multi-User Dungeons (MUD)",
]

dependencies = [
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
ai = [
    "openai>=1.0.0",
    "anthropic>=0.7.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/terminal_game"
Documentation = "https://github.com/yourusername/terminal_game#readme"
Repository = "https://github.com/yourusername/terminal_game"
Issues = "https://github.com/yourusername/terminal_game/issues"

# Entry point removido - execute diretamente: python3 server.py
# [project.scripts]
# mud-server = "server:main"

[tool.setuptools]
packages = ["utils"]

[tool.black]
line-length = 100
target-version = ['py37']

[tool.isort]
profile = "black"
line_length = 100

